% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_api.R
\name{apply_move}
\alias{apply_move}
\title{Apply a move to a board state}
\usage{
apply_move(state, move)
}
\arguments{
\item{state}{A \code{bg_board} object.}

\item{move}{A legal \code{bg_move_sequence} object.}
}
\value{
A \code{bg_board} object after applying the move.
}
\description{
Alias for \code{\link[=bg_apply_move_sequence]{bg_apply_move_sequence()}} using action terminology.
}
