% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_research_frontdoor.R
\name{bg_move_features}
\alias{bg_move_features}
\title{Compute move-level structural features}
\usage{
bg_move_features(x)
}
\arguments{
\item{x}{A \code{bg_problem}, \code{bg_move_sequence}, or list of move sequences.}
}
\value{
A data frame of move features.
}
\description{
Compute move-level structural features
}
