% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_player.R
\name{bg_play_turn_defensive_player}
\alias{bg_play_turn_defensive_player}
\title{Play one turn with the defensive archetype}
\usage{
bg_play_turn_defensive_player(board, roll = NULL, seed = NULL)
}
\arguments{
\item{board}{A \code{bg_board} object.}

\item{roll}{Optional \code{bg_roll} object or roll-like list. If \code{NULL}, dice are
rolled in C++.}

\item{seed}{Optional integer-like scalar. This controls dice rolling when
\code{roll = NULL}.}
}
\value{
A \code{bg_turn_result} object.
}
\description{
Convenience wrapper around \code{\link[=bg_play_turn]{bg_play_turn()}} using \code{selection = "defensive"}.
}
