% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bg_plot_board}
\alias{bg_plot_board}
\title{Plot a board (user wrapper)}
\usage{
bg_plot_board(board, ...)
}
\arguments{
\item{board}{A \code{bg_board} object.}

\item{...}{Unused.}
}
\value{
The input board, invisibly.
}
\description{
Convenience wrapper around \code{\link[=plot.bg_board]{plot.bg_board()}}.
}
