% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_methods.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute mean squared error}
\usage{
compute_mse(estimate, truth)
}
\arguments{
\item{estimate}{Numeric vector of estimates.}

\item{truth}{Numeric vector of benchmark truth values.}
}
\value{
Numeric scalar.
}
\description{
Computes the mean squared error between an estimated move-value vector and a
truth vector.
}
