% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_methods.R
\name{compute_simple_regret}
\alias{compute_simple_regret}
\title{Compute simple regret}
\usage{
compute_simple_regret(selected_value, best_value)
}
\arguments{
\item{selected_value}{Numeric vector of selected-action values.}

\item{best_value}{Numeric vector of best-action values.}
}
\value{
Numeric vector of regrets.
}
\description{
Alias of \code{\link[=compute_regret]{compute_regret()}} used in benchmark summaries.
}
