% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_api.R
\name{explain_move_evaluation}
\alias{explain_move_evaluation}
\title{Explain a move-evaluation object}
\usage{
explain_move_evaluation(x)
}
\arguments{
\item{x}{A \code{bg_action_evaluation} or \code{bg_move_recommendation} object.}
}
\value{
A character scalar.
}
\description{
Explain a move-evaluation object
}
