% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_api.R
\name{initialize_board}
\alias{initialize_board}
\title{Initialize a board state}
\usage{
initialize_board(...)
}
\value{
A validated \code{bg_board} object.
}
\description{
Alias for \code{\link[=bg_position]{bg_position()}} in the research-oriented API.
}
