% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{is_bg_benchmark_case}
\alias{is_bg_benchmark_case}
\title{Test whether an object is a benchmark case}
\usage{
is_bg_benchmark_case(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_benchmark_case"}.
}
\description{
Test whether an object is a benchmark case
}
