% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{is_bg_move_evaluation_benchmark}
\alias{is_bg_move_evaluation_benchmark}
\title{Test whether an object is a move-evaluation benchmark result}
\usage{
is_bg_move_evaluation_benchmark(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_move_evaluation_benchmark"}.
}
\description{
Test whether an object is a move-evaluation benchmark result
}
