% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{is_bg_roll}
\alias{is_bg_roll}
\title{Test whether an object is a backgammon roll}
\usage{
is_bg_roll(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_roll"}; otherwise \code{FALSE}.
}
\description{
Test whether an object is a backgammon roll
}
