% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_api.R
\name{plot_board}
\alias{plot_board}
\title{Plot a board state}
\usage{
plot_board(board, ...)
}
\arguments{
\item{board}{A \code{bg_board} object.}

\item{...}{Unused.}
}
\value{
The input board, invisibly.
}
\description{
Alias for \code{\link[=bg_plot_board]{bg_plot_board()}} in the research-oriented API.
}
