% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{print.bg_analysis_report}
\alias{print.bg_analysis_report}
\title{Print an analysis report}
\usage{
\method{print}{bg_analysis_report}(x, n = 10L, ...)
}
\arguments{
\item{x}{A \code{bg_analysis_report} object.}

\item{n}{Integer-like scalar controlling how many ranked actions to print.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an analysis report
}
