% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bg_board}
\alias{print.bg_board}
\title{Print a backgammon board}
\usage{
\method{print}{bg_board}(x, ...)
}
\arguments{
\item{x}{A \code{bg_board} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a backgammon board
}
