% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_s3_methods.R
\name{print.bg_game_result}
\alias{print.bg_game_result}
\title{Print a game result}
\usage{
\method{print}{bg_game_result}(x, ...)
}
\arguments{
\item{x}{A \code{bg_game_result} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a game result
}
