% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_benchmark.R
\name{print.bg_matchup_benchmark}
\alias{print.bg_matchup_benchmark}
\title{Print a matchup benchmark}
\usage{
\method{print}{bg_matchup_benchmark}(x, ...)
}
\arguments{
\item{x}{A \code{bg_matchup_benchmark} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a matchup benchmark
}
