% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_benchmark.R
\name{print.bg_move_evaluation_benchmark}
\alias{print.bg_move_evaluation_benchmark}
\title{Print a move-evaluation benchmark}
\usage{
\method{print}{bg_move_evaluation_benchmark}(x, ...)
}
\arguments{
\item{x}{A \code{bg_move_evaluation_benchmark} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a move-evaluation benchmark
}
