% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.R
\name{print.bg_move_recommendation}
\alias{print.bg_move_recommendation}
\title{Print a move recommendation}
\usage{
\method{print}{bg_move_recommendation}(x, n = 10L, ...)
}
\arguments{
\item{x}{A \code{bg_move_recommendation} object.}

\item{n}{Integer-like scalar controlling how many ranked moves to print.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a move recommendation
}
