% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_s3_methods.R
\name{print.bg_move_sequence}
\alias{print.bg_move_sequence}
\title{Print a backgammon move sequence}
\usage{
\method{print}{bg_move_sequence}(x, ...)
}
\arguments{
\item{x}{A \code{bg_move_sequence} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a backgammon move sequence
}
