% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thompson_research.R
\name{print.bg_reference_certificate}
\alias{print.bg_reference_certificate}
\title{Print a reference-certificate object}
\usage{
\method{print}{bg_reference_certificate}(x, ...)
}
\arguments{
\item{x}{A \code{bg_reference_certificate} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a reference-certificate object
}
