% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_s3_methods.R
\name{print.bg_roll}
\alias{print.bg_roll}
\title{Print a backgammon dice roll}
\usage{
\method{print}{bg_roll}(x, ...)
}
\arguments{
\item{x}{A \code{bg_roll} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a backgammon dice roll
}
