% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bg_turn_result}
\alias{print.bg_turn_result}
\title{Print a turn result}
\usage{
\method{print}{bg_turn_result}(x, ...)
}
\arguments{
\item{x}{A \code{bg_turn_result} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a turn result
}
