% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_api.R
\name{summarize_benchmark_results}
\alias{summarize_benchmark_results}
\title{Summarize benchmark outputs}
\usage{
summarize_benchmark_results(x)
}
\arguments{
\item{x}{A \code{bg_allocation_benchmark} object.}
}
\value{
A summary data frame.
}
\description{
Summarize benchmark outputs
}
