% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{summary.bg_matchup_benchmark}
\alias{summary.bg_matchup_benchmark}
\title{Summarize a matchup benchmark}
\usage{
\method{summary}{bg_matchup_benchmark}(object, ...)
}
\arguments{
\item{object}{A \code{bg_matchup_benchmark} object.}

\item{...}{Unused.}
}
\value{
A one-row summary data frame.
}
\description{
Summarize a matchup benchmark
}
