% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{summary.bg_move_evaluation_benchmark}
\alias{summary.bg_move_evaluation_benchmark}
\title{Summarize a move-evaluation benchmark}
\usage{
\method{summary}{bg_move_evaluation_benchmark}(object, ...)
}
\arguments{
\item{object}{A \code{bg_move_evaluation_benchmark} object.}

\item{...}{Unused.}
}
\value{
A summary data frame with one row per method.
}
\description{
Summarize a move-evaluation benchmark
}
