% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_algorithms.R
\name{bg_study_load}
\alias{bg_study_load}
\title{Load a study object from disk}
\usage{
bg_study_load(path)
}
\arguments{
\item{path}{Path created by \code{bg_study_save()}.}
}
\value{
A deserialized study object.
}
\description{
Load a study object from disk
}
