% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_truth.R
\name{bg_truth_load}
\alias{bg_truth_load}
\title{Load a truth object from disk}
\usage{
bg_truth_load(path)
}
\arguments{
\item{path}{Path created by \code{bg_truth_save()}.}
}
\value{
A \code{bg_truth_state} or \code{bg_truth_battery} object.
}
\description{
Load a truth object from disk
}
