% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_engine_api.R
\name{is_bg_board}
\alias{is_bg_board}
\title{Test whether an object is a backgammon board}
\usage{
is_bg_board(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_board"}; otherwise \code{FALSE}.
}
\description{
Test whether an object is a backgammon board
}
