% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{is_bg_matchup_simulation}
\alias{is_bg_matchup_simulation}
\title{Test whether an object is a matchup-simulation result}
\usage{
is_bg_matchup_simulation(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_matchup_simulation"}.
}
\description{
Test whether an object is a matchup-simulation result
}
