% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{is_bg_move_sequence}
\alias{is_bg_move_sequence}
\title{Test whether an object is a backgammon move sequence}
\usage{
is_bg_move_sequence(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_move_sequence"}; otherwise
\code{FALSE}.
}
\description{
Test whether an object is a backgammon move sequence
}
