% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_engine_api.R
\name{is_bg_move_step}
\alias{is_bg_move_step}
\title{Test whether an object is an atomic backgammon move step}
\usage{
is_bg_move_step(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_move_step"}; otherwise
\code{FALSE}.
}
\description{
Test whether an object is an atomic backgammon move step
}
