% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_engine_api.R
\name{is_bg_turn_result}
\alias{is_bg_turn_result}
\title{Test whether an object is a turn result}
\usage{
is_bg_turn_result(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"bg_turn_result"}.
}
\description{
Test whether an object is a turn result
}
