% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bg_matchup_simulation}
\alias{print.bg_matchup_simulation}
\title{Print a matchup-simulation result}
\usage{
\method{print}{bg_matchup_simulation}(x, ...)
}
\arguments{
\item{x}{A \code{bg_matchup_simulation} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a matchup-simulation result
}
