% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_s3_methods.R
\name{print.bg_move_step}
\alias{print.bg_move_step}
\title{Print a backgammon move step}
\usage{
\method{print}{bg_move_step}(x, ...)
}
\arguments{
\item{x}{A \code{bg_move_step} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a backgammon move step
}
